requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
duplicity==0.8.12.0
dynaconf==3.1.7
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
hvac==0.11.2
idna==2.8
install==1.3.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.6.0
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
mysql-client==0.0.1
mysqlclient==2.0.3
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
shortuuid==1.0.8
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.4.7
wadllib==1.3.3
xkit==0.0.0
